Separated all components in home page to individual component files
