+++ next-app/buildspace-ipfs/components/Memes.js
@@ -1,0 +1,39 @@
+import { useState } from "react";
+
+import Item from "./Item";
+import listUploads from "../utils/getIPFS";
+
+const Memes = () => {
+    const [memes, setMemes] = useState([]);
+
+    useEffect(() => {
+        const getMemes = async () => {
+            const results = await listUploads();
+            console.log("results are:", results);
+            setMemes(results);
+        };
+
+        getMemes();
+    }, []);
+
+    const [isLoading, setLoading] = useState(true);
+
+    useEffect(() => {
+        if (memes.length > 0) {
+            setLoading(false);
+        }
+    }, [memes]);
+
+    const renderMemesContainer = () => {
+        return memes.map((meme, index) => <Item key={index} item={meme} setLoading={setLoading} className="meme" />);
+    };
+
+    return (
+        <section className="memes-container">
+            {isLoading && <span>Loading...</span>}
+            <div className="grid">{renderMemesContainer()}</div>
+        </section>
+    );
+};
+
+export default Memes;
